refactor(results): extracted to components
