+++ src/components/pages/Results.tsx
@@ -1,24 +1,12 @@
-import {
-  Container,
-  Grid,
-  IconButton,
-  InputAdornment,
-  Switch,
-  TextField,
-  Typography,
-} from '@material-ui/core';
-import React, { useEffect, useState } from 'react';
-import { Build, FileCopy } from '@material-ui/icons';
+import { Container, Grid, Typography } from '@material-ui/core';
+import React from 'react';
 import { useLocation } from 'react-router';
 
 import { list as supportedPlugins } from '../../lib/plugins';
 import { addConfigToCollector } from '../../lib/intermediate-config-collector-builder';
 import { IntermediateConfigCollector } from '../../lib/interfaces/intermediate-config-collector';
-import { createPartialFromCollector } from '../../lib/eslint-config-partial-builder';
-import {
-  BuildTools,
-  createNpmOrYarnCommand,
-} from '../../lib/npm-yarn-command-builder';
+import CommandDisplay from '../organisms/CommandDisplay';
+import ConfigDisplay from '../molecules/ConfigDisplay';
 
 const Results = () => {
   const { search } = useLocation();
@@ -36,29 +24,6 @@ const Results = () => {
     null
   );
 
-  // generate the config file
-  const [eslintConfigPartial, setEslintConfigPartial] = useState('');
-  useEffect(() => {
-    if (collector) {
-      const partialEslint = createPartialFromCollector(collector);
-      setEslintConfigPartial(JSON.stringify(partialEslint, null, 2));
-    }
-  }, [collector]);
-
-  // generate command
-  const [commandType, setCommandType] = useState(BuildTools.YARN);
-  const [installCommand, setInstallCommand] = useState('');
-  useEffect(() => {
-    if (collector) {
-      const command = createNpmOrYarnCommand(collector, commandType);
-      setInstallCommand(command);
-    }
-  }, [collector, commandType]);
-
-  const copyToClipboard = (content: string) => {
-    navigator.clipboard.writeText(content);
-  };
-
   return (
     <Container maxWidth="lg">
       <Grid container spacing={6} direction="column">
@@ -70,39 +35,8 @@ const Results = () => {
         <Grid item xs={12}>
           <Typography variant="h6">Install dependencies</Typography>
         </Grid>
-        <Grid container item xs={12} justify="space-between">
-          <Grid item xs={10}>
-            <TextField
-              id="command"
-              fullWidth
-              value={installCommand}
-              InputProps={{
-                readOnly: true,
-                endAdornment: (
-                  <InputAdornment position="end">
-                    <IconButton onClick={() => copyToClipboard(installCommand)}>
-                      <FileCopy />
-                    </IconButton>
-                  </InputAdornment>
-                ),
-              }}
-            />
-          </Grid>
-          <Grid item>
-            <Grid container alignItems="center" spacing={1}>
-              <Grid item>yarn</Grid>
-              <Grid item>
-                <Switch
-                  onChange={(e) =>
-                    e.target.checked
-                      ? setCommandType(BuildTools.NPM)
-                      : setCommandType(BuildTools.YARN)
-                  }
-                />
-              </Grid>
-              <Grid item>npm</Grid>
-            </Grid>
-          </Grid>
+        <Grid item xs={12}>
+          <CommandDisplay collector={collector} />
         </Grid>
         <Grid item xs={12}>
           <Typography variant="h6">
@@ -110,25 +44,7 @@ const Results = () => {
           </Typography>
         </Grid>
         <Grid item xs={12}>
-          <TextField
-            id="config"
-            multiline
-            fullWidth
-            variant="filled"
-            value={eslintConfigPartial}
-            InputProps={{
-              readOnly: true,
-              endAdornment: (
-                <InputAdornment position="end">
-                  <IconButton
-                    onClick={() => copyToClipboard(eslintConfigPartial)}
-                  >
-                    <FileCopy />
-                  </IconButton>
-                </InputAdornment>
-              ),
-            }}
-          />
+          <ConfigDisplay collector={collector} />
         </Grid>
       </Grid>
     </Container>
